# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  model.ArgOSSToken:
    properties:
      file_name:
        description: 文件名
        type: string
      file_type:
        description: |-
          文件类型
          file: 文章附件
          image: 图片
          certification: 工作证件照片
          other: 其他
        type: string
    type: object
...
  model.PolicyToken:
    properties:
      accessid:
        type: string
      callback:
        type: string
      dir:
        type: string
      expire:
        type: string
      host:
        type: string
      key:
        type: string
      policy:
        type: string
      signature:
        type: string
    type: object
...
host: dev.flywk.com
info:
  contact: {}
  description: |+
    飞行百科 API
    所有返回结果以如下JSON对象返回
    <pre>
    {
    "code": 0,           // 如果没有另行约定，一般为200
    "message": "string", // 一般用户发生错误或者验证失败时候返回的消息
    "result": {},        // 返回结果，所有文档所示的返回结果处于这个字段
    "success": true      // 用于判断是否执行成功
    }
    </pre>

  license: {}
  title: 飞行百科 API
  version: "1.0"
paths:
...
  /file/oss_token:
    post:
      consumes:
      - application/json
      description: |-
        获取阿里云OSS上传TOKEN
        阿里云文档：https://help.aliyun.com/document_detail/31926.html
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Source 来源，1:Web, 2:iOS; 3:Android
        enum:
        - 1
        - 2
        - 3
        in: header
        name: Source
        required: true
        type: integer
      - description: 语言
        enum:
        - zh-CN
        - en-US
        in: header
        name: Locale
        required: true
        type: string
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.ArgOSSToken'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            $ref: '#/definitions/model.PolicyToken'
            type: object
        "400":
          description: 验证失败
        "401":
          description: 登录验证失败
        "500":
          description: 服务器端错误
      summary: 获取阿里云OSS上传TOKEN
      tags:
      - common
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"